move the two interfaces in client module to api module
