+++ clients/client-java/src/main/java/com/datastrato/gravitino/client/GravitinoAdminClient.java
@@ -11,7 +11,6 @@ import com.datastrato.gravitino.authorization.Privilege;
 import com.datastrato.gravitino.authorization.Role;
 import com.datastrato.gravitino.authorization.SecurableObject;
 import com.datastrato.gravitino.authorization.User;
-import com.datastrato.gravitino.client.api.SupportsMetalakes;
 import com.datastrato.gravitino.dto.requests.GroupAddRequest;
 import com.datastrato.gravitino.dto.requests.MetalakeCreateRequest;
 import com.datastrato.gravitino.dto.requests.MetalakeUpdateRequest;
@@ -36,6 +35,7 @@ import com.datastrato.gravitino.exceptions.NoSuchRoleException;
 import com.datastrato.gravitino.exceptions.NoSuchUserException;
 import com.datastrato.gravitino.exceptions.RoleAlreadyExistsException;
 import com.datastrato.gravitino.exceptions.UserAlreadyExistsException;
+import com.datastrato.gravitino.rel.SupportsMetalakes;
 import com.google.common.base.Preconditions;
 import java.util.Arrays;
 import java.util.Collections;
