+++ clients/client-java/src/main/java/com/datastrato/gravitino/client/GravitinoClient.java
@@ -8,10 +8,10 @@ package com.datastrato.gravitino.client;
 import com.datastrato.gravitino.Catalog;
 import com.datastrato.gravitino.CatalogChange;
 import com.datastrato.gravitino.NameIdentifier;
-import com.datastrato.gravitino.client.api.SupportsCatalogs;
 import com.datastrato.gravitino.exceptions.CatalogAlreadyExistsException;
 import com.datastrato.gravitino.exceptions.NoSuchCatalogException;
 import com.datastrato.gravitino.exceptions.NoSuchMetalakeException;
+import com.datastrato.gravitino.rel.SupportsCatalogs;
 import com.google.common.base.Preconditions;
 import java.util.Map;
