+++ clients/client-java/src/main/java/com/datastrato/gravitino/client/GravitinoMetalake.java
@@ -7,7 +7,6 @@ package com.datastrato.gravitino.client;
 import com.datastrato.gravitino.Catalog;
 import com.datastrato.gravitino.CatalogChange;
 import com.datastrato.gravitino.NameIdentifier;
-import com.datastrato.gravitino.client.api.SupportsCatalogs;
 import com.datastrato.gravitino.dto.AuditDTO;
 import com.datastrato.gravitino.dto.MetalakeDTO;
 import com.datastrato.gravitino.dto.requests.CatalogCreateRequest;
@@ -20,6 +19,7 @@ import com.datastrato.gravitino.dto.responses.EntityListResponse;
 import com.datastrato.gravitino.exceptions.CatalogAlreadyExistsException;
 import com.datastrato.gravitino.exceptions.NoSuchCatalogException;
 import com.datastrato.gravitino.exceptions.NoSuchMetalakeException;
+import com.datastrato.gravitino.rel.SupportsCatalogs;
 import com.google.common.base.Preconditions;
 import java.util.Arrays;
 import java.util.Collections;
